Rename "VideoPlayer" folder, namespace more types
