+++ Resources/Private/JavaScript/DlfMediaPlayer/controls/ControlPanelButton.js
