+++ Resources/Private/JavaScript/DlfMediaPlayer/controls/ControlPanelButton.test.js
