+++ Resources/Private/JavaScript/DlfMediaPlayer/controls/FlatSeekBar.js
