+++ Resources/Private/JavaScript/DlfMediaPlayer/controls/PresentationTimeTracker.js
