+++ Resources/Private/JavaScript/DlfMediaPlayer/lib/buildTimeString.js
